+++ packages/core/sagas/componentDataSaga.js
@@ -6,7 +6,7 @@ import {
 import fetchComponentData from 'services/fetchComponentData';
 import getLogService from '@irvingjs/services/logService';
 
-const debug = getLogService('irving:sagas:componentData');
+const log = getLogService('irving:sagas:componentData');
 
 export default function* watchComponentData(action) {
   const {
@@ -23,6 +23,6 @@ export default function* watchComponentData(action) {
     }
   } catch (err) {
     yield put(actionReceiveComponentDataError(endpoint, err));
-    yield call(debug.error, err);
+    yield call(log.error, '%o', err);
   }
 }