+++ packages/vip-go/services/generateLogInfo.js
@@ -1,0 +1,35 @@
+const { format } = require('util');
+
+module.exports = function generateLogInfo(method, messages) {
+  const firstMessage = messages[0];
+  let message = firstMessage;
+  let info;
+
+  // Format string messages like winston would.
+  if (! (firstMessage instanceof Error)) {
+    const formattedMessage = format(...messages);
+
+    // error logged, but message isn't a string.
+    if ('error' === method) {
+      message = new Error(formattedMessage);
+    } else {
+      message = formattedMessage;
+    }
+  }
+
+  // Construct info object.
+  if (message instanceof Error) {
+    info = {
+      message: message.message,
+      name: message.name,
+      stack: message.stack,
+    };
+  } else {
+    info = { message };
+  }
+
+  return {
+    level: method,
+    ...info,
+  };
+};